requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
Flask==3.0.0
fonttools==4.59.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
mysql-connector-python==9.5.0
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
patsy==1.0.1
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.31.0
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
statsmodels==0.14.5
tabulate==0.9.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.4
xlsxwriter==3.2.9
